requirements.txt
dotenv
google-genai
lancedb
sentence-transformers
langchain-text-splitters
llama-index